Remove dependency on PDB reader from TestBase
